Extract all the types for custom layouts to the `common/` level to avoid duplication for the view service interfaces.
