+++ src/vs/workbench/services/views/browser/viewDescriptorService.ts
@@ -23,9 +23,10 @@ import { IStringDictionary } from 'vs/base/common/collections';
 import { ILogger, ILoggerService } from 'vs/platform/log/common/log';
 import { Lazy } from 'vs/base/common/lazy';
 // --- Start Positron ---
-import { CustomPositronLayoutDescription, layoutDescriptionToViewInfo } from 'vs/workbench/browser/positronCustomViews';
+import { layoutDescriptionToViewInfo } from 'vs/workbench/browser/positronCustomViews';
 import { IPaneCompositePartService } from 'vs/workbench/services/panecomposite/browser/panecomposite';
 import { PaneCompositeBar } from 'vs/workbench/browser/parts/paneCompositeBar';
+import { CustomPositronLayoutDescription } from 'vs/workbench/common/positronCustomViews';
 // --- End Positron ---
 
 interface IViewsCustomizations {
